chore: Move "Now Playing" screen vinyl rotation logic to a hook
